+++ backend/src/main/java/com/a407/back/domain/User.java
@@ -14,7 +14,7 @@ public class User {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "id", updatable = false)
-    private int id;
+    private long id;
 
     @Column(name = "password")
     private String password;
@@ -23,7 +23,7 @@ public class User {
     private String email;
 
     @Builder
-    public User(int id, String password, String email) {
+    public User(long id, String password, String email) {
         this.id = id;
         this.password = password;
         this.email = email;
